improvements to menu display
rather than always showing the cursor at the top of the screen and
scroll (and loop) menu items, move the cursor down until it reaches the
bottom of the screen, then scroll items (with cursor still at the
bottom). After last item is reached, jump back to top of screen and
list.